refactor: unused code reduce complexity
