--- check/check.go
@@ -1,25 +1,0 @@
-// Copyright © 2019 Aqua Security Software Ltd. <info@aquasec.com>
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package check
-
-// NodeType indicates the type of node (domain-controller, member-server).
-type NodeType string
-
-const (
-	// DomainController a domain-controller node
-	DomainController NodeType = "domain-controller"
-	// MemberServer a member-server
-	MemberServer NodeType = "member-server"
-)
